create initial users
created a method that creates 2 users (if they dont exist)
1. Admin named "Bob"
2. User named "Sally"
this was used to test to see if files would successfully save and load, which was SUCCESSFUL. Will keep in project.
ToDo: create an initiate chatbox method that creates a chatbox between Bob and Sally (only if it doesn't already exist), to test saving and loading of object and to ensure objects immutable variables stay consistent.